refactor: extract snake case utility
