+++ custom_components/thessla_green_modbus/utils.py
@@ -1,0 +1,22 @@
+"""Utility helpers for ThesslaGreen Modbus integration."""
+
+from __future__ import annotations
+
+import re
+
+__all__ = ["_to_snake_case"]
+
+
+def _to_snake_case(name: str) -> str:
+    """Convert register names to snake_case."""
+    replacements = {"flowrate": "flow_rate"}
+    for old, new in replacements.items():
+        name = name.replace(old, new)
+    name = re.sub(r"[\s\-/]", "_", name)
+    name = re.sub(r"([a-z0-9])([A-Z])", r"\1_\2", name)
+    name = re.sub(r"(?<=\D)(\d)", r"_\1", name)
+    name = re.sub(r"__+", "_", name)
+    name = name.lower()
+    token_map = {"temp": "temperature"}
+    tokens = [token_map.get(token, token) for token in name.split("_")]
+    return "_".join(tokens)
